GCP/storage/functions/tag/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
cloudevents==1.*
